Sounds gehen auch mit Lautstärken < 1.0
git-svn-id: svn://svn.berlios.de/dsa-hl@2565 4c79e8ff-cfd4-0310-af45-a38c79f83013

+++ rl/trunk/plugins/fmod3driver/src/Fmod3SoundChannel.cpp
@@ -57,7 +57,7 @@ void Fmod3SoundChannel::play()
     }
     setChannel(dynamic_cast<Fmod3Sound*>(getSound())->createChannel());
 
-	int vol;
+	float vol;
 	if (is3d())
 	{
 		vol = getDriver()->getDefaultSoundVolume();
